move everything to _src
