+++ mpi4jax/_src/cython/cuda_runtime_api.pxd
