+++ mpi4jax/_src/cython/mpi_xla_bridge.pxd
