+++ mpi4jax/_src/warn.py
