+++ setup.py
@@ -10,6 +10,9 @@ except ImportError:
 else:
     HAS_CYTHON = True
 
+CYTHON_SUBMODULE_NAME = "mpi4jax._src.cython"
+CYTHON_SUBMODULE_PATH = "mpi4jax/_src/cython"
+
 
 #######
 # Utils
@@ -143,8 +146,8 @@ def get_extensions():
 
     extensions = [
         Extension(
-            name=f"mpi4jax.cython.{mod}",
-            sources=[f"mpi4jax/cython/{mod}.{ext_suffix}"],
+            name=f"{CYTHON_SUBMODULE_NAME}.{mod}",
+            sources=[f"{CYTHON_SUBMODULE_PATH}/{mod}.{ext_suffix}"],
             include_dirs=mpi_info("compile"),
             library_dirs=mpi_info("libdirs"),
             libraries=mpi_info("libs"),
@@ -156,8 +159,8 @@ def get_extensions():
     if cuda_info("compile"):
         extensions.append(
             Extension(
-                name="mpi4jax.cython.mpi_xla_bridge_gpu",
-                sources=[f"mpi4jax/cython/mpi_xla_bridge_gpu.{ext_suffix}"],
+                name=f"{CYTHON_SUBMODULE_NAME}.mpi_xla_bridge_gpu",
+                sources=[f"{CYTHON_SUBMODULE_PATH}/mpi_xla_bridge_gpu.{ext_suffix}"],
                 include_dirs=mpi_info("compile") + cuda_info("compile"),
                 library_dirs=mpi_info("libdirs") + cuda_info("libdirs"),
                 libraries=mpi_info("libs") + cuda_info("libs"),
