+++ tests/test_collective_ops.py
@@ -715,7 +715,7 @@ def test_deadlock_on_exit(tmp_path):
 
 def test_set_debug_logging(capsys):
     from mpi4jax import Allreduce
-    from mpi4jax.cython.mpi_xla_bridge import set_logging
+    from mpi4jax._src.cython.mpi_xla_bridge import set_logging
 
     arr = jnp.ones((3, 2))
     set_logging(True)
